backend/pyproject.toml
[tool.poetry]
name = "liars-bar-a2a"
version = "0.1.0"
description = "A2A Liar's Bar Game with SecondMe OAuth2"
authors = ["Liar's Bar Team <team@liarsbar.local>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.35"
asyncpg = "^0.30.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.28.0"
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
